+++ src/index.js
@@ -1,4 +1,5 @@
 export * from './components'
+export * from './composables'
 export * from './directives'
 export * from './lib'
 export * from './mixins'
@@ -7,6 +8,7 @@ export * from './utils'
 export * from './lib/validation'
 
 import * as components from './components'
+import * as composables from './composables'
 import * as directives from './directives'
 import * as lib from './lib'
 import * as mixins from './mixins'
@@ -16,6 +18,7 @@ import * as validation from './lib/validation'
 
 export default {
   ...components,
+  ...composables,
   ...directives,
   ...lib,
   ...mixins,